refactor(github-issues-runtime-tests): extract artifact workflow module (#1274)
